pythonidae/nats-client/requirements.txt
nats-py==2.2.0
certifi==2022.12.7
